packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-ci-triggers"
version = "0.1.0"
description = "Code/prompt change listeners, CI/CD hooks, and regression gate definitions"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["ci-triggers","ai","minions"]
authors = [
    { name = "Mehdi Nabhani", email = "mehdi@the-mehdi.com" }
]
dependencies = [
    "minions-sdk>=0.2.1",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/mxn2020/minions-ci-triggers"
Repository = "https://github.com/mxn2020/minions-ci-triggers"

[tool.hatch.build.targets.wheel]
packages = ["minions_ci_triggers"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
